Changement de projet pour le concours
