--- examples/DestructionBox/build.lua
@@ -1,15 +1,0 @@
-EXAMPLE.Name = "DestructionBox"
-
-EXAMPLE.Console = true
-
-EXAMPLE.Files = {
-	"main.cpp"
-}
-
-EXAMPLE.Libraries = {
-	"NazaraCore",
-	"NazaraGraphics",
-	"NazaraRenderer",
-	"NazaraUtility",
-	"NazaraSDK"
-}
