--- examples/DestructionBox/main.cpp
@@ -1,5 +1,0 @@
-#include <iostream>
-
-int main() {
-    std::cout << "Hello World!" << std::endl;
-}
