+++ examples/Tetris/build.lua
@@ -1,0 +1,15 @@
+EXAMPLE.Name = "Tetris"
+
+EXAMPLE.Console = true
+
+EXAMPLE.Files = {
+    "main.cpp"
+}
+
+EXAMPLE.Libraries = {
+    "NazaraCore",
+    "NazaraGraphics",
+    "NazaraRenderer",
+    "NazaraUtility",
+    "NazaraSDK"
+}
