+++ examples/Tetris/main.cpp
@@ -1,0 +1,5 @@
+#include <iostream>
+
+int main() {
+    std::cout << "Hello, World!" << std::endl;
+}
